Fix decoding fail owing to character break across buffer boundary
A fixed number of bytes is read on each iteration, and sometimes these will
end half-way through a code point encoding (e.g. in UTF-8, where a code
point can be represented by up to 6 bytes).

We can’t cleanly detect this specific error, so just tack the erroneous
input on to the start of the next read and try again: if there’s an error
starting at position 0, then actually report an error.