Allow rpl to replace regexes
Replace regexes by default. This seems sensible, as it’s the way that
similar tools, e.g. grep, work. Indeed, I had vaguely expected rpl to work
this way for years, but only now went out of my way to investigate why it
did not when I was trying to use it!

The old behaviour is available using the --literal/-l flag.

Deprecate the --escape/-e flag: it has an illogical name, since
it *un*escapes the pattern, and it has little use (grep has no equivalent).

Bump the version to 1.8rc1 so that users are more likely to look at what is
new when they get this version, and make it an rc so that there’s an
opportunity for comment and testing before I make a final release.